university-ai-agent/requirements.txt
faster-whisper==1.0.3
pyaudio==0.2.14
pyttsx3==2.90
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.4
